@@ -1,6 +1,6 @@
 module github.com/denandz/glorp
 
-go 1.15
+go 1.16
 
 require (
 	github.com/fsnotify/fsnotify v1.6.0
